Make module out of status page
