+++ src/utils/index.js
@@ -1,5 +1,7 @@
 helpers = require('./helpers')
-
+const { makeStatusTables } = require('./status')
+ 
 module.exports = {
-    helpers
+    helpers,
+    makeStatusTables
 }
