+++ src/utils/views/base.html
@@ -1,24 +1,14 @@
-<!DOCTYPE html>
+<!doctype html>
 {% block head %}
-<meta charset="utf8">
-
-{% block title %}
+    <meta charset="utf8">
     <title>
-        Plugin Status
+        {% block title %}
+            Plugin Status
+        {% endblock %}
     </title>
-{% endblock %}
-<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-<style>
-</style>
-
 {% endblock %}
 
 <body>
-
-{% block navbar %}
-
-{% endblock %}
-
 {% block content %}
 {% endblock %}
 </body>
